Add "use_ttl" to return fields if ttl is specified.
After wasting time figuring out why ibapi get would not show the ttl of
a record, even though specified with --rFields, I found that this is the
case if the ttl value is inherited.  Though admittedly not very helpful,
add the "use_ttl" flag to the return_fields list if the list contains
ttl.