Fix TXT record "spaceless" data bug.
After creating a simple TXT record with ibapi txt add:
  rice.edu 'airtable-verification=82413a59e9bfc1494c2ebdd53e567337'
I found I could not delete it (or get, or update).  After debugging
I found this was an issue with all "spaceless" data TXT records.

sanitizeRecordData() adds double quotes to TXT record data.
This is needed for data with spaces, and originally I thought it
would not hurt for data without spaces.  But, at least now, it
apparently does so.  I updated sanitizeRecordData() to only quote
the data if it contains one or more spaces.

While debugging this I found additional issues with TXT records
with data > 255 chars.  ibapi could add them but not get them.
I realized that whether via the UI or ibapi, IB seems to un-quote
the second (perhaps last?) string.  For instance, if I enter this
in the UI text box:  "D1" "D2":
    1) if both D1 and D2 are small, IB de-quotes both: D1 D2
    2) if at least D1 is large, IB de-quotes D2 only: "D1" D2
dig gives the same result for either case: "D1" "D2"

I am not addressing this issue at this time.  For now, ibapi can
add such records, but it cannot get, update or delete them them
if the data is specified on the commnand line.